Use int, not byte for inventory size
I can't see any reason to use byte :)

+++ src/main/java/steamcraft/common/tiles/TileArmorEditor.java
@@ -26,7 +26,7 @@ public class TileArmorEditor extends BaseTileWithInventory implements IInventory
 {
 	public TileArmorEditor()
 	{
-		super((byte) 44);
+		super(44);
 	}
 
 	@Override
